DEVPROD-7857: Send Sentry alerts for 422 errors in Spruce
