+++ apps/spruce/package.json
@@ -50,6 +50,7 @@
   },
   "dependencies": {
     "@apollo/client": "3.10.4",
+    "@apollo/server": "^4.11.0",
     "@emotion/css": "11.13.4",
     "@emotion/react": "11.13.3",
     "@emotion/styled": "11.13.0",
