+++ apps/spruce/src/gql/client/link/logGQLErrorsLink.ts
@@ -1,5 +1,6 @@
 import { Operation } from "@apollo/client";
 import { onError } from "@apollo/client/link/error";
+import { ApolloServerErrorCode } from "@apollo/server/errors";
 import { GraphQLError } from "graphql";
 import { reportError } from "utils/errorReporting";
 import { deleteNestedKey } from "utils/object";
@@ -11,7 +12,20 @@ export const reportingFn =
     if (path) {
       fingerprint.push(...path);
     }
-    reportError(new Error(gqlErr.message), {
+
+    const isValidationError =
+      gqlErr.extensions?.code ===
+      ApolloServerErrorCode.GRAPHQL_VALIDATION_FAILED;
+
+    const err = isValidationError
+      ? new GraphQLError(
+          `GraphQL validation error in '${operation.operationName}': ${gqlErr.message}`,
+        )
+      : new Error(
+          `Error occurred in '${operation.operationName}': ${gqlErr.message}`,
+        );
+
+    const sendError = reportError(err, {
       fingerprint,
       tags: { operationName: operation.operationName },
       context: {
@@ -22,7 +36,13 @@ export const reportingFn =
           "REDACTED",
         ),
       },
-    }).warning();
+    });
+
+    if (isValidationError) {
+      sendError.severe();
+    } else {
+      sendError.warning();
+    }
   };
 
 export const logGQLErrorsLink = (secretFields: string[]) =>
