+++ apps/spruce/src/gql/fetch/index.ts
@@ -7,9 +7,8 @@ export const secretFieldsReq: RequestInit = {
     "content-type": "application/json",
   },
   body: JSON.stringify({
-    // @ts-expect-error: FIXME. This comment was added by an automated script.
     operationName: (SECRET_FIELDS.definitions[0] as OperationDefinitionNode)
-      .name.value,
+      ?.name?.value,
     query: SECRET_FIELDS.loc?.source.body,
     variables: {},
   }),
