+++ apps/spruce/src/utils/request.ts
@@ -1,9 +1,6 @@
 import { getUiUrl } from "./environmentVariables";
 import { reportError } from "./errorReporting";
 
-export const shouldLogoutAndRedirect = (statusCode: number) =>
-  statusCode === 401;
-
 export const post = async (url: string, body: unknown) => {
   try {
     const response = await fetch(`${getUiUrl()}${url}`, {
@@ -28,34 +25,3 @@ const getErrorMessage = (response: Response, method: string) => {
 const handleError = (error: string) => {
   reportError(new Error(error)).warning();
 };
-
-export const fetchWithRetry = <T = any>(
-  url: string,
-  options: RequestInit,
-  retries: number = 3,
-  backoff: number = 150,
-): Promise<{ data: T }> =>
-  new Promise((resolve, reject) => {
-    const attemptFetch = (attempt: number): void => {
-      fetch(url, options)
-        .then((res) => {
-          if (res.ok) {
-            return res.json();
-          }
-          reject(
-            new Error(getErrorMessage(res, "GET"), {
-              cause: { statusCode: res.status, message: res.statusText },
-            }),
-          );
-        })
-        .then((data) => resolve(data))
-        .catch((err) => {
-          if (attempt <= retries) {
-            setTimeout(() => attemptFetch(attempt + 1), backoff * attempt);
-          } else {
-            reject(err);
-          }
-        });
-    };
-    attemptFetch(1);
-  });
